Backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
greenlet==3.2.4
kombu==5.5.4
openpyxl==3.1.5
packaging==25.0
playwright==1.55.0
prompt_toolkit==3.0.52
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
